requirements.txt
alembic==0.9.5
beautifulsoup4==4.6.0
blinker==1.4
browser-cookie3==0.6.2
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
DBUtils==1.2
docutils==0.14
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
future==0.16.0
Geohash==1.0
geohash2==1.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
keyring==10.4.0
lxml==3.8.0
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.13.2
opencv-python==3.3.0.10
pbkdf2==1.3
pyaes==1.6.0
PyExecJS==1.4.1
pygame==1.9.3
PyInstaller==3.2.1
pymongo==3.5.1
PyMySQL==0.7.11
pypiwin32==219
pyserial==3.4
python-dateutil==2.6.1
python-editor==1.0.3
pywin32-ctypes==0.1.2
redis==2.10.6
requests==2.18.4
six==1.10.0
SQLAlchemy==1.1.14
urllib3==1.22
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
